+++ src/zocalo/cli/dlq_check.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import argparse
 
 import workflows.transport
@@ -61,7 +63,7 @@ def check_dlq_rabbitmq(
 def run() -> None:
     zc = zocalo.configuration.from_file()
     zc.activate()
-    parser = argparse.ArgumentParser("dlstbx.dlq_check [options]")
+    parser = argparse.ArgumentParser(usage="zocalo.dlq_check [options]")
     parser.add_argument("-?", action="help", help=argparse.SUPPRESS)
     parser.add_argument(
         "-n",