+++ src/zocalo/cli/go.py
@@ -3,6 +3,8 @@
 #   Process a datacollection
 #
 
+from __future__ import annotations
+
 import argparse
 import getpass
 import json